theme: Account for menu bar when scrolling to target
When navigating to a page with a fragment, the browser will scroll
such that the element with that id is at the top of the viewport,
but that's also where the menubar is.

`scroll-padding-top` tells the browser to scroll just a little bit
less, so that the target isn't obscured by the menu bar.