Add `xgettext` command to extract translatable strings
This command is one half of a Gettext-based translation (i18n)
workflow. It iterates over each chapter and extracts all translatable
text into a `messages.pot` file.

The text is split on paragraph boundaries, which helps ensure less
churn in the output when the text is edited.

The other half of the workflow is a `gettext` command which will take
a source Markdown file and a `xx.po` file and output a translated
Markdown file.

Part of the solution for #5.