Add `book.translations`
When set, `build` runs a renderer for each translations after setting
`book.language` to a subdirectory.

For example, when it's set `["ko", "ja"]`, `mdbook build -d book` will
generate the output:

```
book/
  (default output)
  ko/
    (ko output)
  ja/
    (ja output)
```

This enables translation to be handled by preprocessors. For example, it
could be PO-file based translation based on `book.language`.