Add FAQ to the documentation
Allow us to use different words to describe certain features to
make it easier for people to find it using their own vocabulary.
Without the need to sneak in the documentation all the possible words.

e.g. In this case searching for "exclude" could not help the
user as that word was not used in our documentation.

See #2218