Fix to not do more than the original code!
Fix: actually in the original code, only two entries were stripped of
their (potential) quotes: "mdbook-theme" and "mdbook-sidebar". Since
now we store sidebar config in sessionStorage, and it cannot contain
legacy values, only "mdbook-theme" needs to be stripped of its quotes,
and if getConfig() is meant to be used more generally, it must not
silently do that to *every* value it reads.